+++ models/components/updatemedia.go
@@ -406,8 +406,8 @@ type UpdateMedia struct {
 	IsAudioOnly optionalnullable.OptionalNullable[bool] `json:"isAudioOnly,omitzero"`
 	// Specifies whether subtitle tracks are available for the media.
 	SubtitleAvailable *bool `json:"subtitleAvailable,omitzero"`
-	// The length of the media in seconds, with a maximum allowed duration of 12 hours per individual media.
-	Duration *string `json:"duration,omitzero"`
+	// Duration of the media in seconds.
+	Duration *float64 `json:"duration,omitzero"`
 	// The aspect ratio of a video is a value that describes the relative shape of a video based on its width and height.
 	AspectRatio *string `json:"aspectRatio,omitzero"`
 	// Time the media was created, defined as a localDateTime (UTC Time).
@@ -581,7 +581,7 @@ func (u *UpdateMedia) GetSubtitleAvailable() *bool {
 	return u.SubtitleAvailable
 }
 
-func (u *UpdateMedia) GetDuration() *string {
+func (u *UpdateMedia) GetDuration() *float64 {
 	if u == nil {
 		return nil
 	}